requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
tqdm
streamlit
langchain-google-genai
google-generativeai
sentence_transformers
huggingface-hub
scikit-learn
numpy
langchain-anthropic
langchain-groq
